requirements.txt
openai>=1.40.0
pydantic>=2.7.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
rich>=13.7.1
pywinauto>=0.6.8
pygetwindow>=0.0.9
psutil>=5.9.8
pyautogui>=0.9.54
watchdog>=4.0.0
PySide6>=6.7.0
Pillow>=10.3.0
mss>=9.0.1